+++ src/xqt-kotlinx-lsp/commonTest/xqt/kotlinx/lsp/test/textDocument/DocumentFormattingParams.kt
@@ -4,6 +4,7 @@ package xqt.kotlinx.lsp.test.textDocument
 import kotlinx.serialization.json.JsonNull
 import kotlinx.serialization.json.JsonPrimitive
 import xqt.kotlinx.lsp.textDocument.DocumentFormattingParams
+import xqt.kotlinx.lsp.types.DocumentUri
 import xqt.kotlinx.rpc.json.serialization.UnsupportedKindTypeException
 import xqt.kotlinx.rpc.json.serialization.jsonArrayOf
 import xqt.kotlinx.rpc.json.serialization.jsonObjectOf
@@ -28,7 +29,7 @@ class TheTextDocumentFormattingParameters {
         )
 
         val params = DocumentFormattingParams.deserialize(json)
-        assertEquals("file:///home/lorem/ipsum.py", params.textDocument.uri)
+        assertEquals(DocumentUri("file:///home/lorem/ipsum.py"), params.textDocument.uri)
 
         assertEquals(8u, params.options.tabSize)
         assertEquals(true, params.options.insertSpaces)