+++ src/xqt-kotlinx-lsp/commonTest/xqt/kotlinx/lsp/test/types/VersionedTextDocumentIdentifier.kt
@@ -3,6 +3,7 @@ package xqt.kotlinx.lsp.test.types
 
 import kotlinx.serialization.json.JsonNull
 import kotlinx.serialization.json.JsonPrimitive
+import xqt.kotlinx.lsp.types.DocumentUri
 import xqt.kotlinx.lsp.types.VersionedTextDocumentIdentifier
 import xqt.kotlinx.rpc.json.serialization.UnsupportedKindTypeException
 import xqt.kotlinx.rpc.json.serialization.jsonArrayOf
@@ -23,7 +24,7 @@ class TheVersionedTextDocumentIdentifierType {
         )
 
         val tdi = VersionedTextDocumentIdentifier.deserialize(json)
-        assertEquals("file:///home/lorem/ipsum.py", tdi.uri)
+        assertEquals(DocumentUri("file:///home/lorem/ipsum.py"), tdi.uri)
         assertEquals(12, tdi.version)
 
         assertEquals(json, VersionedTextDocumentIdentifier.serializeToJson(tdi))