+++ CHANGELOG.md
@@ -25,6 +25,7 @@ Types:
 Types:
 - `Location` -- Changed the `uri` parameter to a `DocumentUri`.
 - `WorkspaceEdit` -- Changed the `changes` parameter key to a `DocumentUri`.
+- `TextDocumentIdentifier` -- Changed the `uri` parameter to a `DocumentUri`.
 
 ### Dependency Upgrades